requirements.txt
starlette
uvicorn
aiofiles
fastapi
tqdm
python-dotenv
requests
matplotlib
oauthlib
requests-oauthlib
numpy
Pillow